+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 * fixed docs for `GlobalRelativeValue`
 * `MemoryPressureManager` will name it's thread
 * enabled _read_in_file to return None as default value
+* added `CancellableCallback`
+* `CallbackGroup` will now support `CancellableCallback`s as well
+* `MemoryPressureManager` will now use `CancellableCallbacks`
+